Adopt "Restrict" semantics and object sanitizer
The LabelDefinition model was not working because dependency information
was unavailable at compile time. Using strings with the "Restrict"
annotation introduces strings, which sucks, but should be better for
analysis.

A side-effect was the convenience of adding "@Restriction(labelName)"
predicates, which can be class-internal. This also makes it possible to
make sanitization an AutoValue extension (I think).