Define scaffolding for server with injected endpoints
The new model will be to have a server (eventually, a generated server
class) that is just a collection of methods for the available endpoints.
Whether or not a generated server like this needs to be part of Jagger
is debatable, it might make more sense to just have a generated
subcomponent for each endpoint.

Using the preferred "testing" configuration of Dagger components, we can
define three different configurations, each with the corresponding
module. Wiring the modules is the next step, so is not done yet.